refactor: use class name for file name
